+++ js/data.js
@@ -1,0 +1,21 @@
+export const MIN_LIKES = 15;
+export const MAX_LIKES = 200;
+export const MIN_COMMENTS = 0;
+export const MAX_COMMENTS = 30;
+export const MIN_AVATAR_ID = 1;
+export const MAX_AVATAR_ID = 6;
+export const MIN_PHOTO_ID = 1;
+export const MAX_PHOTO_ID = 25;
+export const MIN_COMMENT_ID = 100;
+export const MAX_COMMENT_ID = 999;
+
+export const names = ['Артём', 'Иван', 'Ольга', 'Мария', 'Дмитрий', 'Александр', 'Елена', 'Сергей', 'Анна', 'Наталья'];
+
+export const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
