+++ js/helpers.js
@@ -1,0 +1,1 @@
+export const getRandomNumber = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
